+++ lib/lite/uxid.rb
@@ -4,8 +4,8 @@ require "generators/lite/uxid/install_generator" if defined?(Rails::Generators)
 
 require "lite/uxid/version"
 require "lite/uxid/configuration"
-require "lite/uxid/irreversible"
-require "lite/uxid/reversible"
+require "lite/uxid/base/irreversible"
+require "lite/uxid/base/reversible"
 require "lite/uxid/record/hashid"
 require "lite/uxid/record/nanoid"
 require "lite/uxid/record/ulid"