+++ lib/generators/lite/uxid/templates/install.rb
@@ -8,4 +8,5 @@ Lite::Uxid.configure do |config|
   config.nanoid_size = 21
   config.ulid_charset = "0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZ"
   config.ulid_size = 26
+  config.uuid_version = 4
 end