+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ## [1.5.0] - 2024-09-20
+### Added
+- Added uuid version option
 ### Changed
 - Ruby send perf improvements