+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.5.1] - 2024-09-20
+### Changed
+- Use after commit on create instead of after create
+
 ## [1.5.0] - 2024-09-20
 ### Added
 - Added uuid version option